Add add-event form
*added add-event form
*changed Integer to BigDecimal

+++ src/main/java/com/github/java4wro/dto/EventDTO.java
@@ -3,13 +3,14 @@ package com.github.java4wro.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
 public class EventDTO {
 
-    private Integer eventPrice;
+    private BigDecimal eventPrice;
     private String eventName;
     private String eventDescription;
     private LocalDateTime eventDateTime;
